+++ src/scan/BitWeaverHScanner256.cpp
@@ -37,9 +37,9 @@ void BitWeaverHScanner256::scan(int *input, uint64_t numEntry, int *output, Pred
 
     __m256i *simdinput = (__m256i *) input;
     __m256i *simdoutput = (__m256i *) output;
-    int entryInWord = 64 / (entrySize + 1);
-    int numLongWord = numEntry / entryInWord + (numEntry % entryInWord ? 1 : 0);
-    int numSimd = numLongWord / WORD_IN_SIMD + (numLongWord % WORD_IN_SIMD ? 1 : 0);
+    uint64_t entryInWord = 64 / (entrySize + 1);
+    uint64_t numLongWord = numEntry / entryInWord + (numEntry % entryInWord ? 1 : 0);
+    uint64_t numSimd = numLongWord / WORD_IN_SIMD + (numLongWord % WORD_IN_SIMD ? 1 : 0);
 
     switch (p->getOpr()) {
         case opr_eq:
